refactor(bricks): decompose hooks into multiple files
