+++ bricks/dart_frog_dev_server/hooks/test/pre_gen_test.dart
@@ -1,14 +1,10 @@
-import 'dart:io';
-
 import 'package:dart_frog_gen/dart_frog_gen.dart';
 import 'package:mason/mason.dart';
 import 'package:mocktail/mocktail.dart';
-import 'package:path/path.dart' as path;
 import 'package:test/test.dart';
 
 import '../pre_gen.dart';
-
-class _MockHookContext extends Mock implements HookContext {}
+import '../src/exit_overrides.dart';
 
 class _FakeHookContext extends Fake implements HookContext {
   _FakeHookContext({Logger? logger}) : _logger = logger ?? _MockLogger();
@@ -28,8 +24,6 @@ class _FakeHookContext extends Fake implements HookContext {
 }
 
 class _MockLogger extends Mock implements Logger {}
-
-class _MockRouteConfiguration extends Mock implements RouteConfiguration {}
 
 void main() {
   group('preGen', () {
@@ -254,332 +248,6 @@ void main() {
           'invokeCustomEntrypoint': false,
         }),
       );
-    });
-  });
-
-  group('reportRouteConflicts', () {
-    late HookContext context;
-    late Logger logger;
-    late RouteConfiguration configuration;
-
-    setUp(() {
-      context = _MockHookContext();
-      logger = _MockLogger();
-      configuration = _MockRouteConfiguration();
-
-      when(() => context.logger).thenReturn(logger);
-    });
-
-    test('reports nothing when there are no endpoints', () {
-      when(() => configuration.endpoints).thenReturn({});
-      reportRouteConflicts(context, configuration);
-      verifyNever(() => logger.err(any()));
-    });
-
-    test('reports nothing when there are endpoints and no conflicts', () {
-      when(() => configuration.endpoints).thenReturn({
-        '/': const [
-          RouteFile(
-            name: 'index',
-            path: 'index.dart',
-            route: '/',
-            params: [],
-          ),
-        ],
-        '/hello': const [
-          RouteFile(
-            name: 'hello',
-            path: 'hello.dart',
-            route: '/hello',
-            params: [],
-          )
-        ]
-      });
-      reportRouteConflicts(context, configuration);
-      verifyNever(() => logger.err(any()));
-    });
-
-    test('reports single conflict when there is one endpoint with conflicts',
-        () {
-      when(() => configuration.endpoints).thenReturn({
-        '/': const [
-          RouteFile(
-            name: 'index',
-            path: 'index.dart',
-            route: '/',
-            params: [],
-          ),
-        ],
-        '/hello': const [
-          RouteFile(
-            name: 'hello',
-            path: 'hello.dart',
-            route: '/hello',
-            params: [],
-          ),
-          RouteFile(
-            name: 'hello_index',
-            path: 'hello/index.dart',
-            route: '/',
-            params: [],
-          )
-        ]
-      });
-      reportRouteConflicts(context, configuration);
-      verify(
-        () => logger.err(
-          '''Route conflict detected. ${lightCyan.wrap('routes/hello.dart')} and ${lightCyan.wrap('routes/hello/index.dart')} both resolve to ${lightCyan.wrap('/hello')}.''',
-        ),
-      );
-    });
-
-    test(
-        'reports multiple conflicts '
-        'when there are multiple endpoint with conflicts', () {
-      when(() => configuration.endpoints).thenReturn({
-        '/': const [
-          RouteFile(
-            name: 'index',
-            path: 'index.dart',
-            route: '/',
-            params: [],
-          ),
-        ],
-        '/hello': const [
-          RouteFile(
-            name: 'hello',
-            path: 'hello.dart',
-            route: '/hello',
-            params: [],
-          ),
-          RouteFile(
-            name: 'hello_index',
-            path: 'hello/index.dart',
-            route: '/',
-            params: [],
-          )
-        ],
-        '/echo': const [
-          RouteFile(
-            name: 'echo',
-            path: 'echo.dart',
-            route: '/echo',
-            params: [],
-          ),
-          RouteFile(
-            name: 'echo_index',
-            path: 'echo/index.dart',
-            route: '/',
-            params: [],
-          )
-        ]
-      });
-      reportRouteConflicts(context, configuration);
-      verify(
-        () => logger.err(
-          '''Route conflict detected. ${lightCyan.wrap('routes/hello.dart')} and ${lightCyan.wrap('routes/hello/index.dart')} both resolve to ${lightCyan.wrap('/hello')}.''',
-        ),
-      );
-      verify(
-        () => logger.err(
-          '''Route conflict detected. ${lightCyan.wrap('routes/echo.dart')} and ${lightCyan.wrap('routes/echo/index.dart')} both resolve to ${lightCyan.wrap('/echo')}.''',
-        ),
-      );
-    });
-  });
-
-  group('reportRogueRoutes', () {
-    late HookContext context;
-    late Logger logger;
-    late RouteConfiguration configuration;
-
-    setUp(() {
-      context = _MockHookContext();
-      logger = _MockLogger();
-      configuration = _MockRouteConfiguration();
-
-      when(() => context.logger).thenReturn(logger);
-    });
-
-    test('reports nothing when there are no rogue routes', () {
-      when(() => configuration.rogueRoutes).thenReturn([]);
-      reportRogueRoutes(context, configuration);
-      verifyNever(() => logger.err(any()));
-    });
-
-    test('reports single rogue route', () {
-      when(() => configuration.rogueRoutes).thenReturn(
-        const [
-          RouteFile(
-            name: 'hello',
-            path: 'hello.dart',
-            route: '/hello',
-            params: [],
-          ),
-        ],
-      );
-      reportRogueRoutes(context, configuration);
-      verify(
-        () => logger.err(
-          '''Rogue route detected.${defaultForeground.wrap(' ')}Rename ${lightCyan.wrap('routes/hello.dart')} to ${lightCyan.wrap('routes/hello/index.dart')}.''',
-        ),
-      );
-    });
-
-    test('reports multiple rogue routes', () {
-      when(() => configuration.rogueRoutes).thenReturn(
-        const [
-          RouteFile(
-            name: 'hello',
-            path: 'hello.dart',
-            route: '/hello',
-            params: [],
-          ),
-          RouteFile(
-            name: 'hi',
-            path: 'hi.dart',
-            route: '/hi',
-            params: [],
-          ),
-        ],
-      );
-      reportRogueRoutes(context, configuration);
-      verify(
-        () => logger.err(
-          '''Rogue route detected.${defaultForeground.wrap(' ')}Rename ${lightCyan.wrap('routes/hello.dart')} to ${lightCyan.wrap('routes/hello/index.dart')}.''',
-        ),
-      );
-      verify(
-        () => logger.err(
-          '''Rogue route detected.${defaultForeground.wrap(' ')}Rename ${lightCyan.wrap('routes/hi.dart')} to ${lightCyan.wrap('routes/hi/index.dart')}.''',
-        ),
-      );
-    });
-  });
-
-  group('reportExternalPathDependencies', () {
-    late HookContext context;
-    late Logger logger;
-
-    setUp(() {
-      context = _MockHookContext();
-      logger = _MockLogger();
-
-      when(() => context.logger).thenReturn(logger);
-    });
-
-    test('reports nothing when there are no external path dependencies',
-        () async {
-      final directory = Directory.systemTemp.createTempSync();
-      File(path.join(directory.path, 'pubspec.yaml')).writeAsStringSync(
-        '''
-name: example
-version: 0.1.0
-environment:
-  sdk: ^2.17.0
-dependencies:
-  mason: any
-dev_dependencies:
-  test: any
-''',
-      );
-      await expectLater(
-        reportExternalPathDependencies(context, directory),
-        completes,
-      );
-      verifyNever(() => logger.err(any()));
-      directory.delete(recursive: true).ignore();
-    });
-
-    test('reports when there is a single external path dependency', () async {
-      final directory = Directory.systemTemp.createTempSync();
-      File(path.join(directory.path, 'pubspec.yaml')).writeAsStringSync(
-        '''
-name: example
-version: 0.1.0
-environment:
-  sdk: ^2.17.0
-dependencies:
-  mason: any
-  foo:
-    path: ../../foo
-dev_dependencies:
-  test: any
-''',
-      );
-      await expectLater(
-        reportExternalPathDependencies(context, directory),
-        completes,
-      );
-      verify(
-        () => logger.err('All path dependencies must be within the project.'),
-      ).called(1);
-      verify(
-        () => logger.err('External path dependencies detected:'),
-      ).called(1);
-      verify(
-        () => logger.err('  \u{2022} foo from ../../foo'),
-      ).called(1);
-      directory.delete(recursive: true).ignore();
-    });
-
-    test('reports when there are multiple external path dependencies',
-        () async {
-      final directory = Directory.systemTemp.createTempSync();
-      File(path.join(directory.path, 'pubspec.yaml')).writeAsStringSync(
-        '''
-name: example
-version: 0.1.0
-environment:
-  sdk: ^2.17.0
-dependencies:
-  mason: any
-  foo:
-    path: ../../foo
-dev_dependencies:
-  test: any
-  bar:
-    path: ../../bar
-''',
-      );
-      await expectLater(
-        reportExternalPathDependencies(context, directory),
-        completes,
-      );
-      verify(
-        () => logger.err('All path dependencies must be within the project.'),
-      ).called(1);
-      verify(
-        () => logger.err('External path dependencies detected:'),
-      ).called(1);
-      verify(
-        () => logger.err('  \u{2022} foo from ../../foo'),
-      ).called(1);
-      verify(
-        () => logger.err('  \u{2022} bar from ../../bar'),
-      ).called(1);
-      directory.delete(recursive: true).ignore();
-    });
-  });
-
-  group('ExitOverrides', () {
-    group('runZoned', () {
-      test('uses default exit when not specified', () {
-        ExitOverrides.runZoned(() {
-          final overrides = ExitOverrides.current;
-          expect(overrides!.exit, equals(exit));
-        });
-      });
-
-      test('uses custom exit when specified', () {
-        ExitOverrides.runZoned(
-          () {
-            final overrides = ExitOverrides.current;
-            expect(overrides!.exit, isNot(equals(exit)));
-          },
-          exit: (_) {},
-        );
-      });
     });
   });
 }
