+++ bricks/dart_frog_dev_server/hooks/test/src/exit_overrides_test.dart
@@ -1,0 +1,28 @@
+import 'dart:io';
+
+import 'package:test/test.dart';
+
+import '../../src/exit_overrides.dart';
+
+void main() {
+  group('ExitOverrides', () {
+    group('runZoned', () {
+      test('uses default exit when not specified', () {
+        ExitOverrides.runZoned(() {
+          final overrides = ExitOverrides.current;
+          expect(overrides!.exit, equals(exit));
+        });
+      });
+
+      test('uses custom exit when specified', () {
+        ExitOverrides.runZoned(
+          () {
+            final overrides = ExitOverrides.current;
+            expect(overrides!.exit, isNot(equals(exit)));
+          },
+          exit: (_) {},
+        );
+      });
+    });
+  });
+}
