+++ bricks/dart_frog_prod_server/hooks/src/report_external_path_dependencies.dart
@@ -1,0 +1,39 @@
+import 'dart:io' as io;
+
+import 'package:mason/mason.dart';
+import 'package:path/path.dart' as path;
+import 'package:pubspec_parse/pubspec_parse.dart';
+
+Future<void> reportExternalPathDependencies(
+  HookContext context,
+  io.Directory directory,
+  void Function(int exitCode) exit,
+) async {
+  final pubspec = Pubspec.parse(
+    await io.File(path.join(directory.path, 'pubspec.yaml')).readAsString(),
+  );
+
+  final dependencies = pubspec.dependencies;
+  final devDependencies = pubspec.devDependencies;
+  final pathDependencies = [...dependencies.entries, ...devDependencies.entries]
+      .where((entry) => entry.value is PathDependency)
+      .map((entry) {
+    final value = entry.value as PathDependency;
+    return [entry.key, value.path];
+  }).toList();
+  final externalDependencies = pathDependencies.where(
+    (dep) => !path.isWithin(directory.path, dep.last),
+  );
+
+  if (externalDependencies.isNotEmpty) {
+    context.logger
+      ..err('All path dependencies must be within the project.')
+      ..err('External path dependencies detected:');
+    for (final dependency in externalDependencies) {
+      final dependencyName = dependency.first;
+      final dependencyPath = path.normalize(dependency.last);
+      context.logger.err('  \u{2022} $dependencyName from $dependencyPath');
+    }
+    exit(1);
+  }
+}
