+++ bricks/dart_frog_prod_server/hooks/test/post_gen_test.dart
@@ -1,11 +1,13 @@
 import 'dart:io';
 
-import 'package:mason/mason.dart';
+import 'package:mason/mason.dart'
+    show HookContext, Logger, Progress, ExitCode, lightCyan;
 import 'package:mocktail/mocktail.dart';
 import 'package:path/path.dart' as path;
 import 'package:test/test.dart';
 
 import '../post_gen.dart';
+import '../src/exit_overrides.dart';
 
 class _FakeHookContext extends Fake implements HookContext {
   _FakeHookContext({Logger? logger}) : _logger = logger ?? _MockLogger();
@@ -87,102 +89,6 @@ void main() {
         () => logger.info('${lightCyan.wrap('dart build/bin/server.dart')}'),
       ).called(1);
       verifyNever(() => logger.err(any()));
-    });
-
-    group('dartPubGet', () {
-      test('completes when process succeeds', () async {
-        final exitCalls = <int>[];
-        final result = _MockProcessResult();
-        when(() => result.exitCode).thenReturn(ExitCode.success.code);
-        await dartPubGet(
-          context,
-          workingDirectory: '.',
-          runProcess: (
-            executable,
-            args, {
-            String? workingDirectory,
-            bool? runInShell,
-          }) async {
-            expect(executable, equals('dart'));
-            expect(args, equals(['pub', 'get']));
-            expect(workingDirectory, equals('.'));
-            expect(runInShell, isTrue);
-            return result;
-          },
-          exit: exitCalls.add,
-        );
-        expect(exitCalls, isEmpty);
-        verifyNever(() => logger.err(any()));
-      });
-
-      test('exits when process fails', () async {
-        const error = 'oops something went wrong';
-        final exitCalls = <int>[];
-        final result = _MockProcessResult();
-        final code = ExitCode.software.code;
-        when(() => result.exitCode).thenReturn(code);
-        when(() => result.stderr).thenReturn(error);
-        await dartPubGet(
-          context,
-          workingDirectory: '.',
-          runProcess: (
-            executable,
-            args, {
-            String? workingDirectory,
-            bool? runInShell,
-          }) async {
-            return result;
-          },
-          exit: exitCalls.add,
-        );
-        expect(exitCalls, equals([code]));
-        verify(() => logger.err(error)).called(1);
-      });
-
-      test('exits when ProcessException occurs', () async {
-        const error = 'oops something went wrong';
-        final exitCalls = <int>[];
-        final result = _MockProcessResult();
-        final code = ExitCode.software.code;
-        when(() => result.exitCode).thenReturn(code);
-        when(() => result.stderr).thenReturn(error);
-        await dartPubGet(
-          context,
-          workingDirectory: '.',
-          runProcess: (
-            executable,
-            args, {
-            String? workingDirectory,
-            bool? runInShell,
-          }) async {
-            throw ProcessException('dart', ['pub', 'get'], error, code);
-          },
-          exit: exitCalls.add,
-        );
-        expect(exitCalls, equals([code]));
-        verify(() => logger.err(error)).called(1);
-      });
-    });
-  });
-
-  group('ExitOverrides', () {
-    group('runZoned', () {
-      test('uses default exit when not specified', () {
-        ExitOverrides.runZoned(() {
-          final overrides = ExitOverrides.current;
-          expect(overrides!.exit, equals(exit));
-        });
-      });
-
-      test('uses custom exit when specified', () {
-        ExitOverrides.runZoned(
-          () {
-            final overrides = ExitOverrides.current;
-            expect(overrides!.exit, isNot(equals(exit)));
-          },
-          exit: (_) {},
-        );
-      });
     });
   });
 }
