+++ packages/dart_frog_cli/lib/src/commands/dev/templates/dart_frog_dev_server_bundle.dart
@@ -16,13 +16,67 @@ final dartFrogDevServerBundle = MasonBundle.fromJson(<String, dynamic>{
     {
       "path": "pre_gen.dart",
       "data":
-          "aW1wb3J0ICdkYXJ0OmFzeW5jJzsKaW1wb3J0ICdkYXJ0OmlvJyBhcyBpbzsKCmltcG9ydCAncGFja2FnZTpkYXJ0X2Zyb2dfZ2VuL2RhcnRfZnJvZ19nZW4uZGFydCc7CmltcG9ydCAncGFja2FnZTptYXNvbi9tYXNvbi5kYXJ0JzsKaW1wb3J0ICdwYWNrYWdlOnBhdGgvcGF0aC5kYXJ0JyBhcyBwYXRoOwppbXBvcnQgJ3BhY2thZ2U6cHVic3BlY19wYXJzZS9wdWJzcGVjX3BhcnNlLmRhcnQnOwoKdHlwZWRlZiBSb3V0ZUNvbmZpZ3VyYXRpb25CdWlsZGVyID0gUm91dGVDb25maWd1cmF0aW9uIEZ1bmN0aW9uKAogIGlvLkRpcmVjdG9yeSBkaXJlY3RvcnksCik7Cgp2b2lkIF9kZWZhdWx0RXhpdChpbnQgY29kZSkgPT4gRXhpdE92ZXJyaWRlcy5jdXJyZW50Py5leGl0ID8/IGlvLmV4aXQ7CgpGdXR1cmU8dm9pZD4gcnVuKEhvb2tDb250ZXh0IGNvbnRleHQpIGFzeW5jID0+IHByZUdlbihjb250ZXh0KTsKCkZ1dHVyZTx2b2lkPiBwcmVHZW4oCiAgSG9va0NvbnRleHQgY29udGV4dCwgewogIFJvdXRlQ29uZmlndXJhdGlvbkJ1aWxkZXIgYnVpbGRDb25maWd1cmF0aW9uID0gYnVpbGRSb3V0ZUNvbmZpZ3VyYXRpb24sCiAgdm9pZCBGdW5jdGlvbihpbnQgZXhpdENvZGUpIGV4aXQgPSBfZGVmYXVsdEV4aXQsCn0pIGFzeW5jIHsKICBmaW5hbCBSb3V0ZUNvbmZpZ3VyYXRpb24gY29uZmlndXJhdGlvbjsKICB0cnkgewogICAgY29uZmlndXJhdGlvbiA9IGJ1aWxkQ29uZmlndXJhdGlvbihpby5EaXJlY3RvcnkuY3VycmVudCk7CiAgfSBjYXRjaCAoZXJyb3IpIHsKICAgIGNvbnRleHQubG9nZ2VyLmVycignJGVycm9yJyk7CiAgICByZXR1cm4gZXhpdCgxKTsKICB9CgogIHJlcG9ydFJvdXRlQ29uZmxpY3RzKGNvbnRleHQsIGNvbmZpZ3VyYXRpb24pOwogIHJlcG9ydFJvZ3VlUm91dGVzKGNvbnRleHQsIGNvbmZpZ3VyYXRpb24pOwogIGF3YWl0IHJlcG9ydEV4dGVybmFsUGF0aERlcGVuZGVuY2llcyhjb250ZXh0LCBpby5EaXJlY3RvcnkuY3VycmVudCk7CgogIGNvbnRleHQudmFycyA9IHsKICAgICdwb3J0JzogY29udGV4dC52YXJzWydwb3J0J10gPz8gJzgwODAnLAogICAgJ2RpcmVjdG9yaWVzJzogY29uZmlndXJhdGlvbi5kaXJlY3RvcmllcwogICAgICAgIC5tYXAoKGMpID0+IGMudG9Kc29uKCkpCiAgICAgICAgLnRvTGlzdCgpCiAgICAgICAgLnJldmVyc2VkCiAgICAgICAgLnRvTGlzdCgpLAogICAgJ3JvdXRlcyc6IGNvbmZpZ3VyYXRpb24ucm91dGVzLm1hcCgocikgPT4gci50b0pzb24oKSkudG9MaXN0KCksCiAgICAnbWlkZGxld2FyZSc6IGNvbmZpZ3VyYXRpb24ubWlkZGxld2FyZS5tYXAoKG0pID0+IG0udG9Kc29uKCkpLnRvTGlzdCgpLAogICAgJ2dsb2JhbE1pZGRsZXdhcmUnOiBjb25maWd1cmF0aW9uLmdsb2JhbE1pZGRsZXdhcmUgIT0gbnVsbAogICAgICAgID8gY29uZmlndXJhdGlvbi5nbG9iYWxNaWRkbGV3YXJlIS50b0pzb24oKQogICAgICAgIDogZmFsc2UsCiAgICAnc2VydmVTdGF0aWNGaWxlcyc6IGNvbmZpZ3VyYXRpb24uc2VydmVTdGF0aWNGaWxlcywKICAgICdpbnZva2VDdXN0b21FbnRyeXBvaW50JzogY29uZmlndXJhdGlvbi5pbnZva2VDdXN0b21FbnRyeXBvaW50LAogIH07Cn0KCkZ1dHVyZTx2b2lkPiByZXBvcnRFeHRlcm5hbFBhdGhEZXBlbmRlbmNpZXMoCiAgSG9va0NvbnRleHQgY29udGV4dCwKICBpby5EaXJlY3RvcnkgZGlyZWN0b3J5LAopIGFzeW5jIHsKICBmaW5hbCBwdWJzcGVjID0gUHVic3BlYy5wYXJzZSgKICAgIGF3YWl0IGlvLkZpbGUocGF0aC5qb2luKGRpcmVjdG9yeS5wYXRoLCAncHVic3BlYy55YW1sJykpLnJlYWRBc1N0cmluZygpLAogICk7CgogIGZpbmFsIGRlcGVuZGVuY2llcyA9IHB1YnNwZWMuZGVwZW5kZW5jaWVzOwogIGZpbmFsIGRldkRlcGVuZGVuY2llcyA9IHB1YnNwZWMuZGV2RGVwZW5kZW5jaWVzOwogIGZpbmFsIHBhdGhEZXBlbmRlbmNpZXMgPSBbLi4uZGVwZW5kZW5jaWVzLmVudHJpZXMsIC4uLmRldkRlcGVuZGVuY2llcy5lbnRyaWVzXQogICAgICAud2hlcmUoKGVudHJ5KSA9PiBlbnRyeS52YWx1ZSBpcyBQYXRoRGVwZW5kZW5jeSkKICAgICAgLm1hcCgoZW50cnkpIHsKICAgIGZpbmFsIHZhbHVlID0gZW50cnkudmFsdWUgYXMgUGF0aERlcGVuZGVuY3k7CiAgICByZXR1cm4gW2VudHJ5LmtleSwgdmFsdWUucGF0aF07CiAgfSkudG9MaXN0KCk7CiAgZmluYWwgZXh0ZXJuYWxEZXBlbmRlbmNpZXMgPSBwYXRoRGVwZW5kZW5jaWVzLndoZXJlKAogICAgKGRlcCkgPT4gIXBhdGguaXNXaXRoaW4oZGlyZWN0b3J5LnBhdGgsIGRlcC5sYXN0KSwKICApOwoKICBpZiAoZXh0ZXJuYWxEZXBlbmRlbmNpZXMuaXNOb3RFbXB0eSkgewogICAgY29udGV4dC5sb2dnZXIKICAgICAgLi5pbmZvKCcnKQogICAgICAuLmVycignQWxsIHBhdGggZGVwZW5kZW5jaWVzIG11c3QgYmUgd2l0aGluIHRoZSBwcm9qZWN0LicpCiAgICAgIC4uZXJyKCdFeHRlcm5hbCBwYXRoIGRlcGVuZGVuY2llcyBkZXRlY3RlZDonKTsKICAgIGZvciAoZmluYWwgZGVwZW5kZW5jeSBpbiBleHRlcm5hbERlcGVuZGVuY2llcykgewogICAgICBmaW5hbCBkZXBlbmRlbmN5TmFtZSA9IGRlcGVuZGVuY3kuZmlyc3Q7CiAgICAgIGZpbmFsIGRlcGVuZGVuY3lQYXRoID0gcGF0aC5ub3JtYWxpemUoZGVwZW5kZW5jeS5sYXN0KTsKICAgICAgY29udGV4dC5sb2dnZXIuZXJyKCcgIFx1ezIwMjJ9ICRkZXBlbmRlbmN5TmFtZSBmcm9tICRkZXBlbmRlbmN5UGF0aCcpOwogICAgfQogIH0KfQoKdm9pZCByZXBvcnRSb3V0ZUNvbmZsaWN0cygKICBIb29rQ29udGV4dCBjb250ZXh0LAogIFJvdXRlQ29uZmlndXJhdGlvbiBjb25maWd1cmF0aW9uLAopIHsKICBmaW5hbCBjb25mbGljdGluZ0VuZHBvaW50cyA9CiAgICAgIGNvbmZpZ3VyYXRpb24uZW5kcG9pbnRzLmVudHJpZXMud2hlcmUoKGVudHJ5KSA9PiBlbnRyeS52YWx1ZS5sZW5ndGggPiAxKTsKICBpZiAoY29uZmxpY3RpbmdFbmRwb2ludHMuaXNOb3RFbXB0eSkgewogICAgY29udGV4dC5sb2dnZXIuaW5mbygnJyk7CiAgICBmb3IgKGZpbmFsIGNvbmZsaWN0IGluIGNvbmZsaWN0aW5nRW5kcG9pbnRzKSB7CiAgICAgIGZpbmFsIG9yaWdpbmFsRmlsZVBhdGggPSBwYXRoLm5vcm1hbGl6ZSgKICAgICAgICBwYXRoLmpvaW4oJ3JvdXRlcycsIGNvbmZsaWN0LnZhbHVlLmZpcnN0LnBhdGgpLAogICAgICApOwogICAgICBmaW5hbCBjb25mbGljdGluZ0ZpbGVQYXRoID0gcGF0aC5ub3JtYWxpemUoCiAgICAgICAgcGF0aC5qb2luKCdyb3V0ZXMnLCBjb25mbGljdC52YWx1ZS5sYXN0LnBhdGgpLAogICAgICApOwogICAgICBjb250ZXh0LmxvZ2dlci5lcnIoCiAgICAgICAgJycnUm91dGUgY29uZmxpY3QgZGV0ZWN0ZWQuICR7bGlnaHRDeWFuLndyYXAob3JpZ2luYWxGaWxlUGF0aCl9IGFuZCAke2xpZ2h0Q3lhbi53cmFwKGNvbmZsaWN0aW5nRmlsZVBhdGgpfSBib3RoIHJlc29sdmUgdG8gJHtsaWdodEN5YW4ud3JhcChjb25mbGljdC5rZXkpfS4nJycsCiAgICAgICk7CiAgICB9CiAgfQp9Cgp2b2lkIHJlcG9ydFJvZ3VlUm91dGVzKAogIEhvb2tDb250ZXh0IGNvbnRleHQsCiAgUm91dGVDb25maWd1cmF0aW9uIGNvbmZpZ3VyYXRpb24sCikgewogIGlmIChjb25maWd1cmF0aW9uLnJvZ3VlUm91dGVzLmlzTm90RW1wdHkpIHsKICAgIGNvbnRleHQubG9nZ2VyLmluZm8oJycpOwogICAgZm9yIChmaW5hbCByb3V0ZSBpbiBjb25maWd1cmF0aW9uLnJvZ3VlUm91dGVzKSB7CiAgICAgIGZpbmFsIGZpbGVQYXRoID0gcGF0aC5ub3JtYWxpemUocGF0aC5qb2luKCdyb3V0ZXMnLCByb3V0ZS5wYXRoKSk7CiAgICAgIGZpbmFsIGZpbGVEaXJlY3RvcnkgPSBwYXRoLmRpcm5hbWUoZmlsZVBhdGgpOwogICAgICBmaW5hbCBpZGVhbFBhdGggPSBwYXRoLmpvaW4oCiAgICAgICAgZmlsZURpcmVjdG9yeSwKICAgICAgICBwYXRoLmJhc2VuYW1lV2l0aG91dEV4dGVuc2lvbihmaWxlUGF0aCksCiAgICAgICAgJ2luZGV4LmRhcnQnLAogICAgICApOwogICAgICBjb250ZXh0LmxvZ2dlci5lcnIoCiAgICAgICAgJycnUm9ndWUgcm91dGUgZGV0ZWN0ZWQuJHtkZWZhdWx0Rm9yZWdyb3VuZC53cmFwKCcgJyl9UmVuYW1lICR7bGlnaHRDeWFuLndyYXAoZmlsZVBhdGgpfSB0byAke2xpZ2h0Q3lhbi53cmFwKGlkZWFsUGF0aCl9LicnJywKICAgICAgKTsKICAgIH0KICB9Cn0KCmNvbnN0IF9hc3luY1J1blpvbmVkID0gcnVuWm9uZWQ7CgphYnN0cmFjdCBjbGFzcyBFeGl0T3ZlcnJpZGVzIHsKICBzdGF0aWMgZmluYWwgX3Rva2VuID0gT2JqZWN0KCk7CgogIHN0YXRpYyBFeGl0T3ZlcnJpZGVzPyBnZXQgY3VycmVudCB7CiAgICByZXR1cm4gWm9uZS5jdXJyZW50W190b2tlbl0gYXMgRXhpdE92ZXJyaWRlcz87CiAgfQoKICBzdGF0aWMgUiBydW5ab25lZDxSPihSIEZ1bmN0aW9uKCkgYm9keSwge3ZvaWQgRnVuY3Rpb24oaW50KT8gZXhpdH0pIHsKICAgIGZpbmFsIG92ZXJyaWRlcyA9IF9FeGl0T3ZlcnJpZGVzU2NvcGUoZXhpdCk7CiAgICByZXR1cm4gX2FzeW5jUnVuWm9uZWQoYm9keSwgem9uZVZhbHVlczoge190b2tlbjogb3ZlcnJpZGVzfSk7CiAgfQoKICB2b2lkIEZ1bmN0aW9uKGludCBleGl0Q29kZSkgZ2V0IGV4aXQgPT4gaW8uZXhpdDsKfQoKY2xhc3MgX0V4aXRPdmVycmlkZXNTY29wZSBleHRlbmRzIEV4aXRPdmVycmlkZXMgewogIF9FeGl0T3ZlcnJpZGVzU2NvcGUodGhpcy5fZXhpdCk7CgogIGZpbmFsIEV4aXRPdmVycmlkZXM/IF9wcmV2aW91cyA9IEV4aXRPdmVycmlkZXMuY3VycmVudDsKICBmaW5hbCB2b2lkIEZ1bmN0aW9uKGludCBleGl0Q29kZSk/IF9leGl0OwoKICBAb3ZlcnJpZGUKICB2b2lkIEZ1bmN0aW9uKGludCBleGl0Q29kZSkgZ2V0IGV4aXQgewogICAgcmV0dXJuIF9leGl0ID8/IF9wcmV2aW91cz8uZXhpdCA/PyBzdXBlci5leGl0OwogIH0KfQo=",
+          "aW1wb3J0ICdkYXJ0OmFzeW5jJzsKaW1wb3J0ICdkYXJ0OmlvJyBhcyBpbzsKCmltcG9ydCAncGFja2FnZTpkYXJ0X2Zyb2dfZ2VuL2RhcnRfZnJvZ19nZW4uZGFydCc7CmltcG9ydCAncGFja2FnZTptYXNvbi9tYXNvbi5kYXJ0JyBzaG93IEhvb2tDb250ZXh0OwoKaW1wb3J0ICdzcmMvZXhpdF9vdmVycmlkZXMuZGFydCc7CmltcG9ydCAnc3JjL3JlcG9ydF9leHRlcm5hbF9wYXRoX2RlcGVuZGVuY2llcy5kYXJ0JzsKaW1wb3J0ICdzcmMvcmVwb3J0X3JvZ3VlX3JvdXRlcy5kYXJ0JzsKaW1wb3J0ICdzcmMvcmVwb3J0X3JvdXRlX2NvbmZsaWN0cy5kYXJ0JzsKCnR5cGVkZWYgUm91dGVDb25maWd1cmF0aW9uQnVpbGRlciA9IFJvdXRlQ29uZmlndXJhdGlvbiBGdW5jdGlvbigKICBpby5EaXJlY3RvcnkgZGlyZWN0b3J5LAopOwoKdm9pZCBfZGVmYXVsdEV4aXQoaW50IGNvZGUpID0+IEV4aXRPdmVycmlkZXMuY3VycmVudD8uZXhpdCA/PyBpby5leGl0OwoKRnV0dXJlPHZvaWQ+IHJ1bihIb29rQ29udGV4dCBjb250ZXh0KSBhc3luYyA9PiBwcmVHZW4oY29udGV4dCk7CgpGdXR1cmU8dm9pZD4gcHJlR2VuKAogIEhvb2tDb250ZXh0IGNvbnRleHQsIHsKICBSb3V0ZUNvbmZpZ3VyYXRpb25CdWlsZGVyIGJ1aWxkQ29uZmlndXJhdGlvbiA9IGJ1aWxkUm91dGVDb25maWd1cmF0aW9uLAogIHZvaWQgRnVuY3Rpb24oaW50IGV4aXRDb2RlKSBleGl0ID0gX2RlZmF1bHRFeGl0LAp9KSBhc3luYyB7CiAgZmluYWwgUm91dGVDb25maWd1cmF0aW9uIGNvbmZpZ3VyYXRpb247CiAgdHJ5IHsKICAgIGNvbmZpZ3VyYXRpb24gPSBidWlsZENvbmZpZ3VyYXRpb24oaW8uRGlyZWN0b3J5LmN1cnJlbnQpOwogIH0gY2F0Y2ggKGVycm9yKSB7CiAgICBjb250ZXh0LmxvZ2dlci5lcnIoJyRlcnJvcicpOwogICAgcmV0dXJuIGV4aXQoMSk7CiAgfQoKICByZXBvcnRSb3V0ZUNvbmZsaWN0cyhjb250ZXh0LCBjb25maWd1cmF0aW9uKTsKICByZXBvcnRSb2d1ZVJvdXRlcyhjb250ZXh0LCBjb25maWd1cmF0aW9uKTsKICBhd2FpdCByZXBvcnRFeHRlcm5hbFBhdGhEZXBlbmRlbmNpZXMoY29udGV4dCwgaW8uRGlyZWN0b3J5LmN1cnJlbnQpOwoKICBjb250ZXh0LnZhcnMgPSB7CiAgICAncG9ydCc6IGNvbnRleHQudmFyc1sncG9ydCddID8/ICc4MDgwJywKICAgICdkaXJlY3Rvcmllcyc6IGNvbmZpZ3VyYXRpb24uZGlyZWN0b3JpZXMKICAgICAgICAubWFwKChjKSA9PiBjLnRvSnNvbigpKQogICAgICAgIC50b0xpc3QoKQogICAgICAgIC5yZXZlcnNlZAogICAgICAgIC50b0xpc3QoKSwKICAgICdyb3V0ZXMnOiBjb25maWd1cmF0aW9uLnJvdXRlcy5tYXAoKHIpID0+IHIudG9Kc29uKCkpLnRvTGlzdCgpLAogICAgJ21pZGRsZXdhcmUnOiBjb25maWd1cmF0aW9uLm1pZGRsZXdhcmUubWFwKChtKSA9PiBtLnRvSnNvbigpKS50b0xpc3QoKSwKICAgICdnbG9iYWxNaWRkbGV3YXJlJzogY29uZmlndXJhdGlvbi5nbG9iYWxNaWRkbGV3YXJlICE9IG51bGwKICAgICAgICA/IGNvbmZpZ3VyYXRpb24uZ2xvYmFsTWlkZGxld2FyZSEudG9Kc29uKCkKICAgICAgICA6IGZhbHNlLAogICAgJ3NlcnZlU3RhdGljRmlsZXMnOiBjb25maWd1cmF0aW9uLnNlcnZlU3RhdGljRmlsZXMsCiAgICAnaW52b2tlQ3VzdG9tRW50cnlwb2ludCc6IGNvbmZpZ3VyYXRpb24uaW52b2tlQ3VzdG9tRW50cnlwb2ludCwKICB9Owp9Cg==",
       "type": "text"
     },
     {
       "path": "pubspec.yaml",
       "data":
           "bmFtZTogZGFydF9mcm9nX2Rldl9zZXJ2ZXJfaG9va3MKcHVibGlzaF90bzogbm9uZQoKZW52aXJvbm1lbnQ6CiAgc2RrOiAiPj0yLjE4LjAgPDMuMC4wIgoKZGVwZW5kZW5jaWVzOgogIGRhcnRfZnJvZ19nZW46IF4wLjMuMAogIG1hc29uOiBeMC4xLjAtZGV2LjM4CiAgcHVic3BlY19wYXJzZTogXjEuMi4wCgpkZXZfZGVwZW5kZW5jaWVzOgogIG1vY2t0YWlsOiBeMC4zLjAKICBwYXRoOiBeMS44LjIKICB0ZXN0OiBeMS4xOS4yCiAgdmVyeV9nb29kX2FuYWx5c2lzOiBeMy4xLjAK",
+      "type": "text"
+    },
+    {
+      "path": "src/exit_overrides.dart",
+      "data":
+          "aW1wb3J0ICdkYXJ0OmFzeW5jJzsKaW1wb3J0ICdkYXJ0OmlvJyBhcyBpbzsKCmNvbnN0IF9hc3luY1J1blpvbmVkID0gcnVuWm9uZWQ7CgphYnN0cmFjdCBjbGFzcyBFeGl0T3ZlcnJpZGVzIHsKICBzdGF0aWMgZmluYWwgX3Rva2VuID0gT2JqZWN0KCk7CgogIHN0YXRpYyBFeGl0T3ZlcnJpZGVzPyBnZXQgY3VycmVudCB7CiAgICByZXR1cm4gWm9uZS5jdXJyZW50W190b2tlbl0gYXMgRXhpdE92ZXJyaWRlcz87CiAgfQoKICBzdGF0aWMgUiBydW5ab25lZDxSPihSIEZ1bmN0aW9uKCkgYm9keSwge3ZvaWQgRnVuY3Rpb24oaW50KT8gZXhpdH0pIHsKICAgIGZpbmFsIG92ZXJyaWRlcyA9IF9FeGl0T3ZlcnJpZGVzU2NvcGUoZXhpdCk7CiAgICByZXR1cm4gX2FzeW5jUnVuWm9uZWQoYm9keSwgem9uZVZhbHVlczoge190b2tlbjogb3ZlcnJpZGVzfSk7CiAgfQoKICB2b2lkIEZ1bmN0aW9uKGludCBleGl0Q29kZSkgZ2V0IGV4aXQgPT4gaW8uZXhpdDsKfQoKY2xhc3MgX0V4aXRPdmVycmlkZXNTY29wZSBleHRlbmRzIEV4aXRPdmVycmlkZXMgewogIF9FeGl0T3ZlcnJpZGVzU2NvcGUodGhpcy5fZXhpdCk7CgogIGZpbmFsIEV4aXRPdmVycmlkZXM/IF9wcmV2aW91cyA9IEV4aXRPdmVycmlkZXMuY3VycmVudDsKICBmaW5hbCB2b2lkIEZ1bmN0aW9uKGludCBleGl0Q29kZSk/IF9leGl0OwoKICBAb3ZlcnJpZGUKICB2b2lkIEZ1bmN0aW9uKGludCBleGl0Q29kZSkgZ2V0IGV4aXQgewogICAgcmV0dXJuIF9leGl0ID8/IF9wcmV2aW91cz8uZXhpdCA/PyBzdXBlci5leGl0OwogIH0KfQo=",
+      "type": "text"
+    },
+    {
+      "path": "src/report_external_path_dependencies.dart",
+      "data":
+          "aW1wb3J0ICdkYXJ0OmlvJyBhcyBpbzsKCmltcG9ydCAncGFja2FnZTptYXNvbi9tYXNvbi5kYXJ0JzsKaW1wb3J0ICdwYWNrYWdlOnBhdGgvcGF0aC5kYXJ0JyBhcyBwYXRoOwppbXBvcnQgJ3BhY2thZ2U6cHVic3BlY19wYXJzZS9wdWJzcGVjX3BhcnNlLmRhcnQnOwoKRnV0dXJlPHZvaWQ+IHJlcG9ydEV4dGVybmFsUGF0aERlcGVuZGVuY2llcygKICBIb29rQ29udGV4dCBjb250ZXh0LAogIGlvLkRpcmVjdG9yeSBkaXJlY3RvcnksCikgYXN5bmMgewogIGZpbmFsIHB1YnNwZWMgPSBQdWJzcGVjLnBhcnNlKAogICAgYXdhaXQgaW8uRmlsZShwYXRoLmpvaW4oZGlyZWN0b3J5LnBhdGgsICdwdWJzcGVjLnlhbWwnKSkucmVhZEFzU3RyaW5nKCksCiAgKTsKCiAgZmluYWwgZGVwZW5kZW5jaWVzID0gcHVic3BlYy5kZXBlbmRlbmNpZXM7CiAgZmluYWwgZGV2RGVwZW5kZW5jaWVzID0gcHVic3BlYy5kZXZEZXBlbmRlbmNpZXM7CiAgZmluYWwgcGF0aERlcGVuZGVuY2llcyA9IFsuLi5kZXBlbmRlbmNpZXMuZW50cmllcywgLi4uZGV2RGVwZW5kZW5jaWVzLmVudHJpZXNdCiAgICAgIC53aGVyZSgoZW50cnkpID0+IGVudHJ5LnZhbHVlIGlzIFBhdGhEZXBlbmRlbmN5KQogICAgICAubWFwKChlbnRyeSkgewogICAgZmluYWwgdmFsdWUgPSBlbnRyeS52YWx1ZSBhcyBQYXRoRGVwZW5kZW5jeTsKICAgIHJldHVybiBbZW50cnkua2V5LCB2YWx1ZS5wYXRoXTsKICB9KS50b0xpc3QoKTsKICBmaW5hbCBleHRlcm5hbERlcGVuZGVuY2llcyA9IHBhdGhEZXBlbmRlbmNpZXMud2hlcmUoCiAgICAoZGVwKSA9PiAhcGF0aC5pc1dpdGhpbihkaXJlY3RvcnkucGF0aCwgZGVwLmxhc3QpLAogICk7CgogIGlmIChleHRlcm5hbERlcGVuZGVuY2llcy5pc05vdEVtcHR5KSB7CiAgICBjb250ZXh0LmxvZ2dlcgogICAgICAuLmluZm8oJycpCiAgICAgIC4uZXJyKCdBbGwgcGF0aCBkZXBlbmRlbmNpZXMgbXVzdCBiZSB3aXRoaW4gdGhlIHByb2plY3QuJykKICAgICAgLi5lcnIoJ0V4dGVybmFsIHBhdGggZGVwZW5kZW5jaWVzIGRldGVjdGVkOicpOwogICAgZm9yIChmaW5hbCBkZXBlbmRlbmN5IGluIGV4dGVybmFsRGVwZW5kZW5jaWVzKSB7CiAgICAgIGZpbmFsIGRlcGVuZGVuY3lOYW1lID0gZGVwZW5kZW5jeS5maXJzdDsKICAgICAgZmluYWwgZGVwZW5kZW5jeVBhdGggPSBwYXRoLm5vcm1hbGl6ZShkZXBlbmRlbmN5Lmxhc3QpOwogICAgICBjb250ZXh0LmxvZ2dlci5lcnIoJyAgXHV7MjAyMn0gJGRlcGVuZGVuY3lOYW1lIGZyb20gJGRlcGVuZGVuY3lQYXRoJyk7CiAgICB9CiAgfQp9Cg==",
+      "type": "text"
+    },
+    {
+      "path": "src/report_rogue_routes.dart",
+      "data":
+          "aW1wb3J0ICdwYWNrYWdlOmRhcnRfZnJvZ19nZW4vZGFydF9mcm9nX2dlbi5kYXJ0JzsKaW1wb3J0ICdwYWNrYWdlOm1hc29uL21hc29uLmRhcnQnOwppbXBvcnQgJ3BhY2thZ2U6cGF0aC9wYXRoLmRhcnQnIGFzIHBhdGg7Cgp2b2lkIHJlcG9ydFJvZ3VlUm91dGVzKAogIEhvb2tDb250ZXh0IGNvbnRleHQsCiAgUm91dGVDb25maWd1cmF0aW9uIGNvbmZpZ3VyYXRpb24sCikgewogIGlmIChjb25maWd1cmF0aW9uLnJvZ3VlUm91dGVzLmlzTm90RW1wdHkpIHsKICAgIGNvbnRleHQubG9nZ2VyLmluZm8oJycpOwogICAgZm9yIChmaW5hbCByb3V0ZSBpbiBjb25maWd1cmF0aW9uLnJvZ3VlUm91dGVzKSB7CiAgICAgIGZpbmFsIGZpbGVQYXRoID0gcGF0aC5ub3JtYWxpemUocGF0aC5qb2luKCdyb3V0ZXMnLCByb3V0ZS5wYXRoKSk7CiAgICAgIGZpbmFsIGZpbGVEaXJlY3RvcnkgPSBwYXRoLmRpcm5hbWUoZmlsZVBhdGgpOwogICAgICBmaW5hbCBpZGVhbFBhdGggPSBwYXRoLmpvaW4oCiAgICAgICAgZmlsZURpcmVjdG9yeSwKICAgICAgICBwYXRoLmJhc2VuYW1lV2l0aG91dEV4dGVuc2lvbihmaWxlUGF0aCksCiAgICAgICAgJ2luZGV4LmRhcnQnLAogICAgICApOwogICAgICBjb250ZXh0LmxvZ2dlci5lcnIoCiAgICAgICAgJycnUm9ndWUgcm91dGUgZGV0ZWN0ZWQuJHtkZWZhdWx0Rm9yZWdyb3VuZC53cmFwKCcgJyl9UmVuYW1lICR7bGlnaHRDeWFuLndyYXAoZmlsZVBhdGgpfSB0byAke2xpZ2h0Q3lhbi53cmFwKGlkZWFsUGF0aCl9LicnJywKICAgICAgKTsKICAgIH0KICB9Cn0K",
+      "type": "text"
+    },
+    {
+      "path": "src/report_route_conflicts.dart",
+      "data":
+          "aW1wb3J0ICdwYWNrYWdlOmRhcnRfZnJvZ19nZW4vZGFydF9mcm9nX2dlbi5kYXJ0JzsKaW1wb3J0ICdwYWNrYWdlOm1hc29uL21hc29uLmRhcnQnOwppbXBvcnQgJ3BhY2thZ2U6cGF0aC9wYXRoLmRhcnQnIGFzIHBhdGg7Cgp2b2lkIHJlcG9ydFJvdXRlQ29uZmxpY3RzKAogIEhvb2tDb250ZXh0IGNvbnRleHQsCiAgUm91dGVDb25maWd1cmF0aW9uIGNvbmZpZ3VyYXRpb24sCikgewogIGZpbmFsIGNvbmZsaWN0aW5nRW5kcG9pbnRzID0KICAgICAgY29uZmlndXJhdGlvbi5lbmRwb2ludHMuZW50cmllcy53aGVyZSgoZW50cnkpID0+IGVudHJ5LnZhbHVlLmxlbmd0aCA+IDEpOwogIGlmIChjb25mbGljdGluZ0VuZHBvaW50cy5pc05vdEVtcHR5KSB7CiAgICBjb250ZXh0LmxvZ2dlci5pbmZvKCcnKTsKICAgIGZvciAoZmluYWwgY29uZmxpY3QgaW4gY29uZmxpY3RpbmdFbmRwb2ludHMpIHsKICAgICAgZmluYWwgb3JpZ2luYWxGaWxlUGF0aCA9IHBhdGgubm9ybWFsaXplKAogICAgICAgIHBhdGguam9pbigncm91dGVzJywgY29uZmxpY3QudmFsdWUuZmlyc3QucGF0aCksCiAgICAgICk7CiAgICAgIGZpbmFsIGNvbmZsaWN0aW5nRmlsZVBhdGggPSBwYXRoLm5vcm1hbGl6ZSgKICAgICAgICBwYXRoLmpvaW4oJ3JvdXRlcycsIGNvbmZsaWN0LnZhbHVlLmxhc3QucGF0aCksCiAgICAgICk7CiAgICAgIGNvbnRleHQubG9nZ2VyLmVycigKICAgICAgICAnJydSb3V0ZSBjb25mbGljdCBkZXRlY3RlZC4gJHtsaWdodEN5YW4ud3JhcChvcmlnaW5hbEZpbGVQYXRoKX0gYW5kICR7bGlnaHRDeWFuLndyYXAoY29uZmxpY3RpbmdGaWxlUGF0aCl9IGJvdGggcmVzb2x2ZSB0byAke2xpZ2h0Q3lhbi53cmFwKGNvbmZsaWN0LmtleSl9LicnJywKICAgICAgKTsKICAgIH0KICB9Cn0K",
+      "type": "text"
+    },
+    {
+      "path": "test/pre_gen_test.dart",
+      "data":
+          "aW1wb3J0ICdwYWNrYWdlOmRhcnRfZnJvZ19nZW4vZGFydF9mcm9nX2dlbi5kYXJ0JzsKaW1wb3J0ICdwYWNrYWdlOm1hc29uL21hc29uLmRhcnQnOwppbXBvcnQgJ3BhY2thZ2U6bW9ja3RhaWwvbW9ja3RhaWwuZGFydCc7CmltcG9ydCAncGFja2FnZTp0ZXN0L3Rlc3QuZGFydCc7CgppbXBvcnQgJy4uL3ByZV9nZW4uZGFydCc7CmltcG9ydCAnLi4vc3JjL2V4aXRfb3ZlcnJpZGVzLmRhcnQnOwoKY2xhc3MgX0Zha2VIb29rQ29udGV4dCBleHRlbmRzIEZha2UgaW1wbGVtZW50cyBIb29rQ29udGV4dCB7CiAgX0Zha2VIb29rQ29udGV4dCh7TG9nZ2VyPyBsb2dnZXJ9KSA6IF9sb2dnZXIgPSBsb2dnZXIgPz8gX01vY2tMb2dnZXIoKTsKCiAgZmluYWwgTG9nZ2VyIF9sb2dnZXI7CgogIHZhciBfdmFycyA9IDxTdHJpbmcsIGR5bmFtaWM+e307CgogIEBvdmVycmlkZQogIE1hcDxTdHJpbmcsIGR5bmFtaWM+IGdldCB2YXJzID0+IF92YXJzOwoKICBAb3ZlcnJpZGUKICBzZXQgdmFycyhNYXA8U3RyaW5nLCBkeW5hbWljPiB2YWx1ZSkgPT4gX3ZhcnMgPSB2YWx1ZTsKCiAgQG92ZXJyaWRlCiAgTG9nZ2VyIGdldCBsb2dnZXIgPT4gX2xvZ2dlcjsKfQoKY2xhc3MgX01vY2tMb2dnZXIgZXh0ZW5kcyBNb2NrIGltcGxlbWVudHMgTG9nZ2VyIHt9Cgp2b2lkIG1haW4oKSB7CiAgZ3JvdXAoJ3ByZUdlbicsICgpIHsKICAgIGxhdGUgSG9va0NvbnRleHQgY29udGV4dDsKICAgIGxhdGUgTG9nZ2VyIGxvZ2dlcjsKCiAgICBzZXRVcCgoKSB7CiAgICAgIGxvZ2dlciA9IF9Nb2NrTG9nZ2VyKCk7CiAgICAgIGNvbnRleHQgPSBfRmFrZUhvb2tDb250ZXh0KGxvZ2dlcjogbG9nZ2VyKTsKICAgIH0pOwoKICAgIHRlc3QoJ3J1biBjb21wbGV0ZXMnLCAoKSB7CiAgICAgIGV4cGVjdCgKICAgICAgICBFeGl0T3ZlcnJpZGVzLnJ1blpvbmVkKAogICAgICAgICAgKCkgPT4gcnVuKF9GYWtlSG9va0NvbnRleHQoKSksCiAgICAgICAgICBleGl0OiAoXykge30sCiAgICAgICAgKSwKICAgICAgICBjb21wbGV0ZXMsCiAgICAgICk7CiAgICB9KTsKCiAgICB0ZXN0KCdleGl0KDEpIGlmIGJ1aWxkUm91dGVDb25maWd1cmF0aW9uIHRocm93cycsICgpIGFzeW5jIHsKICAgICAgZmluYWwgZXhpdENhbGxzID0gPGludD5bXTsKICAgICAgZmluYWwgZXhjZXB0aW9uID0gRXhjZXB0aW9uKCdvb3BzJyk7CiAgICAgIGF3YWl0IHByZUdlbigKICAgICAgICBjb250ZXh0LAogICAgICAgIGJ1aWxkQ29uZmlndXJhdGlvbjogKF8pID0+IHRocm93IGV4Y2VwdGlvbiwKICAgICAgICBleGl0OiBleGl0Q2FsbHMuYWRkLAogICAgICApOwogICAgICBleHBlY3QoZXhpdENhbGxzLCBlcXVhbHMoWzFdKSk7CiAgICAgIHZlcmlmeSgoKSA9PiBsb2dnZXIuZXJyKGV4Y2VwdGlvbi50b1N0cmluZygpKSkuY2FsbGVkKDEpOwogICAgfSk7CgogICAgdGVzdCgncmV0YWlucyBjdXN0b20gcG9ydCBpZiBzcGVjaWZpZWQnLCAoKSBhc3luYyB7CiAgICAgIGNvbnN0IGN1c3RvbVBvcnQgPSAnODA4MSc7CiAgICAgIGNvbnRleHQudmFyc1sncG9ydCddID0gY3VzdG9tUG9ydDsKICAgICAgY29uc3QgY29uZmlndXJhdGlvbiA9IFJvdXRlQ29uZmlndXJhdGlvbigKICAgICAgICBtaWRkbGV3YXJlOiBbXSwKICAgICAgICBkaXJlY3RvcmllczogW10sCiAgICAgICAgcm91dGVzOiBbXSwKICAgICAgICByb2d1ZVJvdXRlczogW10sCiAgICAgICAgZW5kcG9pbnRzOiB7fSwKICAgICAgKTsKICAgICAgZmluYWwgZXhpdENhbGxzID0gPGludD5bXTsKICAgICAgYXdhaXQgcHJlR2VuKAogICAgICAgIGNvbnRleHQsCiAgICAgICAgYnVpbGRDb25maWd1cmF0aW9uOiAoXykgPT4gY29uZmlndXJhdGlvbiwKICAgICAgICBleGl0OiBleGl0Q2FsbHMuYWRkLAogICAgICApOwogICAgICBleHBlY3QoZXhpdENhbGxzLCBpc0VtcHR5KTsKICAgICAgdmVyaWZ5TmV2ZXIoKCkgPT4gbG9nZ2VyLmVycihhbnkoKSkpOwogICAgICBleHBlY3QoCiAgICAgICAgY29udGV4dC52YXJzLAogICAgICAgIGVxdWFscyh7CiAgICAgICAgICAncG9ydCc6IGN1c3RvbVBvcnQsCiAgICAgICAgICAnZGlyZWN0b3JpZXMnOiA8Um91dGVEaXJlY3Rvcnk+W10sCiAgICAgICAgICAncm91dGVzJzogPFJvdXRlRmlsZT5bXSwKICAgICAgICAgICdtaWRkbGV3YXJlJzogPE1pZGRsZXdhcmVGaWxlPltdLAogICAgICAgICAgJ2dsb2JhbE1pZGRsZXdhcmUnOiBmYWxzZSwKICAgICAgICAgICdzZXJ2ZVN0YXRpY0ZpbGVzJzogZmFsc2UsCiAgICAgICAgICAnaW52b2tlQ3VzdG9tRW50cnlwb2ludCc6IGZhbHNlCiAgICAgICAgfSksCiAgICAgICk7CiAgICB9KTsKCiAgICB0ZXN0KCdyZXRhaW5zIGludm9rZUN1c3RvbUVudHJ5cG9pbnQgKHRydWUpJywgKCkgYXN5bmMgewogICAgICBjb25zdCBjdXN0b21Qb3J0ID0gJzgwODEnOwogICAgICBjb250ZXh0LnZhcnNbJ3BvcnQnXSA9IGN1c3RvbVBvcnQ7CiAgICAgIGNvbnN0IGNvbmZpZ3VyYXRpb24gPSBSb3V0ZUNvbmZpZ3VyYXRpb24oCiAgICAgICAgbWlkZGxld2FyZTogW10sCiAgICAgICAgZGlyZWN0b3JpZXM6IFtdLAogICAgICAgIHJvdXRlczogW10sCiAgICAgICAgcm9ndWVSb3V0ZXM6IFtdLAogICAgICAgIGVuZHBvaW50czoge30sCiAgICAgICAgaW52b2tlQ3VzdG9tRW50cnlwb2ludDogdHJ1ZSwKICAgICAgKTsKICAgICAgZmluYWwgZXhpdENhbGxzID0gPGludD5bXTsKICAgICAgYXdhaXQgcHJlR2VuKAogICAgICAgIGNvbnRleHQsCiAgICAgICAgYnVpbGRDb25maWd1cmF0aW9uOiAoXykgPT4gY29uZmlndXJhdGlvbiwKICAgICAgICBleGl0OiBleGl0Q2FsbHMuYWRkLAogICAgICApOwogICAgICBleHBlY3QoZXhpdENhbGxzLCBpc0VtcHR5KTsKICAgICAgdmVyaWZ5TmV2ZXIoKCkgPT4gbG9nZ2VyLmVycihhbnkoKSkpOwogICAgICBleHBlY3QoCiAgICAgICAgY29udGV4dC52YXJzLAogICAgICAgIGVxdWFscyh7CiAgICAgICAgICAncG9ydCc6IGN1c3RvbVBvcnQsCiAgICAgICAgICAnZGlyZWN0b3JpZXMnOiA8Um91dGVEaXJlY3Rvcnk+W10sCiAgICAgICAgICAncm91dGVzJzogPFJvdXRlRmlsZT5bXSwKICAgICAgICAgICdtaWRkbGV3YXJlJzogPE1pZGRsZXdhcmVGaWxlPltdLAogICAgICAgICAgJ2dsb2JhbE1pZGRsZXdhcmUnOiBmYWxzZSwKICAgICAgICAgICdzZXJ2ZVN0YXRpY0ZpbGVzJzogZmFsc2UsCiAgICAgICAgICAnaW52b2tlQ3VzdG9tRW50cnlwb2ludCc6IHRydWUsCiAgICAgICAgfSksCiAgICAgICk7CiAgICB9KTsKCiAgICB0ZXN0KCd1cGRhdGVzIGNvbnRleHQudmFycyB3aGVuIGJ1aWxkUm91dGVDb25maWd1cmF0aW9uIHN1Y2NlZWRzJywKICAgICAgICAoKSBhc3luYyB7CiAgICAgIGNvbnN0IGNvbmZpZ3VyYXRpb24gPSBSb3V0ZUNvbmZpZ3VyYXRpb24oCiAgICAgICAgZ2xvYmFsTWlkZGxld2FyZTogTWlkZGxld2FyZUZpbGUoCiAgICAgICAgICBuYW1lOiAnbWlkZGxld2FyZScsCiAgICAgICAgICBwYXRoOiAnbWlkZGxld2FyZS5kYXJ0JywKICAgICAgICApLAogICAgICAgIG1pZGRsZXdhcmU6IFsKICAgICAgICAgIE1pZGRsZXdhcmVGaWxlKAogICAgICAgICAgICBuYW1lOiAnaGVsbG9fbWlkZGxld2FyZScsCiAgICAgICAgICAgIHBhdGg6ICdoZWxsby9taWRkbGV3YXJlLmRhcnQnLAogICAgICAgICAgKQogICAgICAgIF0sCiAgICAgICAgZGlyZWN0b3JpZXM6IFsKICAgICAgICAgIFJvdXRlRGlyZWN0b3J5KAogICAgICAgICAgICBuYW1lOiAnXycsCiAgICAgICAgICAgIHJvdXRlOiAnLycsCiAgICAgICAgICAgIG1pZGRsZXdhcmU6IFtdLAogICAgICAgICAgICBmaWxlczogWwogICAgICAgICAgICAgIFJvdXRlRmlsZSgKICAgICAgICAgICAgICAgIG5hbWU6ICdpbmRleCcsCiAgICAgICAgICAgICAgICBwYXRoOiAnaW5kZXguZGFydCcsCiAgICAgICAgICAgICAgICByb3V0ZTogJy8nLAogICAgICAgICAgICAgICAgcGFyYW1zOiBbXSwKICAgICAgICAgICAgICApLAogICAgICAgICAgICAgIFJvdXRlRmlsZSgKICAgICAgICAgICAgICAgIG5hbWU6ICdoZWxsbycsCiAgICAgICAgICAgICAgICBwYXRoOiAnaGVsbG8uZGFydCcsCiAgICAgICAgICAgICAgICByb3V0ZTogJy9oZWxsbycsCiAgICAgICAgICAgICAgICBwYXJhbXM6IFtdLAogICAgICAgICAgICAgICksCiAgICAgICAgICAgIF0sCiAgICAgICAgICAgIHBhcmFtczogW10sCiAgICAgICAgICApCiAgICAgICAgXSwKICAgICAgICByb3V0ZXM6IFsKICAgICAgICAgIFJvdXRlRmlsZSgKICAgICAgICAgICAgbmFtZTogJ2luZGV4JywKICAgICAgICAgICAgcGF0aDogJ2luZGV4LmRhcnQnLAogICAgICAgICAgICByb3V0ZTogJy8nLAogICAgICAgICAgICBwYXJhbXM6IFtdLAogICAgICAgICAgKSwKICAgICAgICAgIFJvdXRlRmlsZSgKICAgICAgICAgICAgbmFtZTogJ2hlbGxvJywKICAgICAgICAgICAgcGF0aDogJ2hlbGxvLmRhcnQnLAogICAgICAgICAgICByb3V0ZTogJy9oZWxsbycsCiAgICAgICAgICAgIHBhcmFtczogW10sCiAgICAgICAgICApLAogICAgICAgIF0sCiAgICAgICAgZW5kcG9pbnRzOiB7CiAgICAgICAgICAnLyc6IFsKICAgICAgICAgICAgUm91dGVGaWxlKAogICAgICAgICAgICAgIG5hbWU6ICdpbmRleCcsCiAgICAgICAgICAgICAgcGF0aDogJ2luZGV4LmRhcnQnLAogICAgICAgICAgICAgIHJvdXRlOiAnLycsCiAgICAgICAgICAgICAgcGFyYW1zOiBbXSwKICAgICAgICAgICAgKSwKICAgICAgICAgIF0sCiAgICAgICAgICAnL2hlbGxvJzogWwogICAgICAgICAgICBSb3V0ZUZpbGUoCiAgICAgICAgICAgICAgbmFtZTogJ2hlbGxvJywKICAgICAgICAgICAgICBwYXRoOiAnaGVsbG8uZGFydCcsCiAgICAgICAgICAgICAgcm91dGU6ICcvaGVsbG8nLAogICAgICAgICAgICAgIHBhcmFtczogW10sCiAgICAgICAgICAgICksCiAgICAgICAgICBdCiAgICAgICAgfSwKICAgICAgICByb2d1ZVJvdXRlczogW10sCiAgICAgICAgc2VydmVTdGF0aWNGaWxlczogdHJ1ZSwKICAgICAgKTsKICAgICAgZmluYWwgZXhpdENhbGxzID0gPGludD5bXTsKICAgICAgYXdhaXQgcHJlR2VuKAogICAgICAgIGNvbnRleHQsCiAgICAgICAgYnVpbGRDb25maWd1cmF0aW9uOiAoXykgPT4gY29uZmlndXJhdGlvbiwKICAgICAgICBleGl0OiBleGl0Q2FsbHMuYWRkLAogICAgICApOwogICAgICBleHBlY3QoZXhpdENhbGxzLCBpc0VtcHR5KTsKICAgICAgdmVyaWZ5TmV2ZXIoKCkgPT4gbG9nZ2VyLmVycihhbnkoKSkpOwogICAgICBleHBlY3QoCiAgICAgICAgY29udGV4dC52YXJzLAogICAgICAgIGVxdWFscyh7CiAgICAgICAgICAncG9ydCc6ICc4MDgwJywKICAgICAgICAgICdkaXJlY3Rvcmllcyc6IFsKICAgICAgICAgICAgewogICAgICAgICAgICAgICduYW1lJzogJ18nLAogICAgICAgICAgICAgICdyb3V0ZSc6ICcvJywKICAgICAgICAgICAgICAnbWlkZGxld2FyZSc6IDxNYXA8U3RyaW5nLCBkeW5hbWljPj5bXSwKICAgICAgICAgICAgICAnZmlsZXMnOiBbCiAgICAgICAgICAgICAgICB7CiAgICAgICAgICAgICAgICAgICduYW1lJzogJ2luZGV4JywKICAgICAgICAgICAgICAgICAgJ3BhdGgnOiAnaW5kZXguZGFydCcsCiAgICAgICAgICAgICAgICAgICdyb3V0ZSc6ICcvJywKICAgICAgICAgICAgICAgICAgJ2ZpbGVfcGFyYW1zJzogPFN0cmluZz5bXSwKICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICB7CiAgICAgICAgICAgICAgICAgICduYW1lJzogJ2hlbGxvJywKICAgICAgICAgICAgICAgICAgJ3BhdGgnOiAnaGVsbG8uZGFydCcsCiAgICAgICAgICAgICAgICAgICdyb3V0ZSc6ICcvaGVsbG8nLAogICAgICAgICAgICAgICAgICAnZmlsZV9wYXJhbXMnOiA8U3RyaW5nPltdLAogICAgICAgICAgICAgICAgfQogICAgICAgICAgICAgIF0sCiAgICAgICAgICAgICAgJ2RpcmVjdG9yeV9wYXJhbXMnOiA8U3RyaW5nPltdLAogICAgICAgICAgICB9CiAgICAgICAgICBdLAogICAgICAgICAgJ3JvdXRlcyc6IFsKICAgICAgICAgICAgewogICAgICAgICAgICAgICduYW1lJzogJ2luZGV4JywKICAgICAgICAgICAgICAncGF0aCc6ICdpbmRleC5kYXJ0JywKICAgICAgICAgICAgICAncm91dGUnOiAnLycsCiAgICAgICAgICAgICAgJ2ZpbGVfcGFyYW1zJzogY29uc3QgPFN0cmluZz5bXSwKICAgICAgICAgICAgfSwKICAgICAgICAgICAgewogICAgICAgICAgICAgICduYW1lJzogJ2hlbGxvJywKICAgICAgICAgICAgICAncGF0aCc6ICdoZWxsby5kYXJ0JywKICAgICAgICAgICAgICAncm91dGUnOiAnL2hlbGxvJywKICAgICAgICAgICAgICAnZmlsZV9wYXJhbXMnOiBjb25zdCA8U3RyaW5nPltdLAogICAgICAgICAgICB9CiAgICAgICAgICBdLAogICAgICAgICAgJ21pZGRsZXdhcmUnOiBbCiAgICAgICAgICAgIHsnbmFtZSc6ICdoZWxsb19taWRkbGV3YXJlJywgJ3BhdGgnOiAnaGVsbG8vbWlkZGxld2FyZS5kYXJ0J30KICAgICAgICAgIF0sCiAgICAgICAgICAnZ2xvYmFsTWlkZGxld2FyZSc6IHsnbmFtZSc6ICdtaWRkbGV3YXJlJywgJ3BhdGgnOiAnbWlkZGxld2FyZS5kYXJ0J30sCiAgICAgICAgICAnc2VydmVTdGF0aWNGaWxlcyc6IHRydWUsCiAgICAgICAgICAnaW52b2tlQ3VzdG9tRW50cnlwb2ludCc6IGZhbHNlLAogICAgICAgIH0pLAogICAgICApOwogICAgfSk7CiAgfSk7Cn0K",
+      "type": "text"
+    },
+    {
+      "path": "test/src/exit_overrides_test.dart",
+      "data":
+          "aW1wb3J0ICdkYXJ0OmlvJzsKCmltcG9ydCAncGFja2FnZTp0ZXN0L3Rlc3QuZGFydCc7CgppbXBvcnQgJy4uLy4uL3NyYy9leGl0X292ZXJyaWRlcy5kYXJ0JzsKCnZvaWQgbWFpbigpIHsKICBncm91cCgnRXhpdE92ZXJyaWRlcycsICgpIHsKICAgIGdyb3VwKCdydW5ab25lZCcsICgpIHsKICAgICAgdGVzdCgndXNlcyBkZWZhdWx0IGV4aXQgd2hlbiBub3Qgc3BlY2lmaWVkJywgKCkgewogICAgICAgIEV4aXRPdmVycmlkZXMucnVuWm9uZWQoKCkgewogICAgICAgICAgZmluYWwgb3ZlcnJpZGVzID0gRXhpdE92ZXJyaWRlcy5jdXJyZW50OwogICAgICAgICAgZXhwZWN0KG92ZXJyaWRlcyEuZXhpdCwgZXF1YWxzKGV4aXQpKTsKICAgICAgICB9KTsKICAgICAgfSk7CgogICAgICB0ZXN0KCd1c2VzIGN1c3RvbSBleGl0IHdoZW4gc3BlY2lmaWVkJywgKCkgewogICAgICAgIEV4aXRPdmVycmlkZXMucnVuWm9uZWQoCiAgICAgICAgICAoKSB7CiAgICAgICAgICAgIGZpbmFsIG92ZXJyaWRlcyA9IEV4aXRPdmVycmlkZXMuY3VycmVudDsKICAgICAgICAgICAgZXhwZWN0KG92ZXJyaWRlcyEuZXhpdCwgaXNOb3QoZXF1YWxzKGV4aXQpKSk7CiAgICAgICAgICB9LAogICAgICAgICAgZXhpdDogKF8pIHt9LAogICAgICAgICk7CiAgICAgIH0pOwogICAgfSk7CiAgfSk7Cn0K",
+      "type": "text"
+    },
+    {
+      "path": "test/src/report_external_path_dependencies_test.dart",
+      "data":
+          "aW1wb3J0ICdkYXJ0OmlvJzsKCmltcG9ydCAncGFja2FnZTptYXNvbi9tYXNvbi5kYXJ0JzsKaW1wb3J0ICdwYWNrYWdlOm1vY2t0YWlsL21vY2t0YWlsLmRhcnQnOwppbXBvcnQgJ3BhY2thZ2U6cGF0aC9wYXRoLmRhcnQnIGFzIHBhdGg7CmltcG9ydCAncGFja2FnZTp0ZXN0L3Rlc3QuZGFydCc7CgppbXBvcnQgJy4uLy4uL3NyYy9yZXBvcnRfZXh0ZXJuYWxfcGF0aF9kZXBlbmRlbmNpZXMuZGFydCc7CgpjbGFzcyBfTW9ja0hvb2tDb250ZXh0IGV4dGVuZHMgTW9jayBpbXBsZW1lbnRzIEhvb2tDb250ZXh0IHt9CgpjbGFzcyBfTW9ja0xvZ2dlciBleHRlbmRzIE1vY2sgaW1wbGVtZW50cyBMb2dnZXIge30KCnZvaWQgbWFpbigpIHsKICBncm91cCgncmVwb3J0RXh0ZXJuYWxQYXRoRGVwZW5kZW5jaWVzJywgKCkgewogICAgbGF0ZSBIb29rQ29udGV4dCBjb250ZXh0OwogICAgbGF0ZSBMb2dnZXIgbG9nZ2VyOwoKICAgIHNldFVwKCgpIHsKICAgICAgY29udGV4dCA9IF9Nb2NrSG9va0NvbnRleHQoKTsKICAgICAgbG9nZ2VyID0gX01vY2tMb2dnZXIoKTsKCiAgICAgIHdoZW4oKCkgPT4gY29udGV4dC5sb2dnZXIpLnRoZW5SZXR1cm4obG9nZ2VyKTsKICAgIH0pOwoKICAgIHRlc3QoJ3JlcG9ydHMgbm90aGluZyB3aGVuIHRoZXJlIGFyZSBubyBleHRlcm5hbCBwYXRoIGRlcGVuZGVuY2llcycsCiAgICAgICAgKCkgYXN5bmMgewogICAgICBmaW5hbCBkaXJlY3RvcnkgPSBEaXJlY3Rvcnkuc3lzdGVtVGVtcC5jcmVhdGVUZW1wU3luYygpOwogICAgICBGaWxlKHBhdGguam9pbihkaXJlY3RvcnkucGF0aCwgJ3B1YnNwZWMueWFtbCcpKS53cml0ZUFzU3RyaW5nU3luYygKICAgICAgICAnJycKbmFtZTogZXhhbXBsZQp2ZXJzaW9uOiAwLjEuMAplbnZpcm9ubWVudDoKICBzZGs6IF4yLjE3LjAKZGVwZW5kZW5jaWVzOgogIG1hc29uOiBhbnkKZGV2X2RlcGVuZGVuY2llczoKICB0ZXN0OiBhbnkKJycnLAogICAgICApOwogICAgICBhd2FpdCBleHBlY3RMYXRlcigKICAgICAgICByZXBvcnRFeHRlcm5hbFBhdGhEZXBlbmRlbmNpZXMoY29udGV4dCwgZGlyZWN0b3J5KSwKICAgICAgICBjb21wbGV0ZXMsCiAgICAgICk7CiAgICAgIHZlcmlmeU5ldmVyKCgpID0+IGxvZ2dlci5lcnIoYW55KCkpKTsKICAgICAgZGlyZWN0b3J5LmRlbGV0ZShyZWN1cnNpdmU6IHRydWUpLmlnbm9yZSgpOwogICAgfSk7CgogICAgdGVzdCgncmVwb3J0cyB3aGVuIHRoZXJlIGlzIGEgc2luZ2xlIGV4dGVybmFsIHBhdGggZGVwZW5kZW5jeScsICgpIGFzeW5jIHsKICAgICAgZmluYWwgZGlyZWN0b3J5ID0gRGlyZWN0b3J5LnN5c3RlbVRlbXAuY3JlYXRlVGVtcFN5bmMoKTsKICAgICAgRmlsZShwYXRoLmpvaW4oZGlyZWN0b3J5LnBhdGgsICdwdWJzcGVjLnlhbWwnKSkud3JpdGVBc1N0cmluZ1N5bmMoCiAgICAgICAgJycnCm5hbWU6IGV4YW1wbGUKdmVyc2lvbjogMC4xLjAKZW52aXJvbm1lbnQ6CiAgc2RrOiBeMi4xNy4wCmRlcGVuZGVuY2llczoKICBtYXNvbjogYW55CiAgZm9vOgogICAgcGF0aDogLi4vLi4vZm9vCmRldl9kZXBlbmRlbmNpZXM6CiAgdGVzdDogYW55CicnJywKICAgICAgKTsKICAgICAgYXdhaXQgZXhwZWN0TGF0ZXIoCiAgICAgICAgcmVwb3J0RXh0ZXJuYWxQYXRoRGVwZW5kZW5jaWVzKGNvbnRleHQsIGRpcmVjdG9yeSksCiAgICAgICAgY29tcGxldGVzLAogICAgICApOwogICAgICB2ZXJpZnkoCiAgICAgICAgKCkgPT4gbG9nZ2VyLmVycignQWxsIHBhdGggZGVwZW5kZW5jaWVzIG11c3QgYmUgd2l0aGluIHRoZSBwcm9qZWN0LicpLAogICAgICApLmNhbGxlZCgxKTsKICAgICAgdmVyaWZ5KAogICAgICAgICgpID0+IGxvZ2dlci5lcnIoJ0V4dGVybmFsIHBhdGggZGVwZW5kZW5jaWVzIGRldGVjdGVkOicpLAogICAgICApLmNhbGxlZCgxKTsKICAgICAgdmVyaWZ5KAogICAgICAgICgpID0+IGxvZ2dlci5lcnIoJyAgXHV7MjAyMn0gZm9vIGZyb20gLi4vLi4vZm9vJyksCiAgICAgICkuY2FsbGVkKDEpOwogICAgICBkaXJlY3RvcnkuZGVsZXRlKHJlY3Vyc2l2ZTogdHJ1ZSkuaWdub3JlKCk7CiAgICB9KTsKCiAgICB0ZXN0KCdyZXBvcnRzIHdoZW4gdGhlcmUgYXJlIG11bHRpcGxlIGV4dGVybmFsIHBhdGggZGVwZW5kZW5jaWVzJywKICAgICAgICAoKSBhc3luYyB7CiAgICAgIGZpbmFsIGRpcmVjdG9yeSA9IERpcmVjdG9yeS5zeXN0ZW1UZW1wLmNyZWF0ZVRlbXBTeW5jKCk7CiAgICAgIEZpbGUocGF0aC5qb2luKGRpcmVjdG9yeS5wYXRoLCAncHVic3BlYy55YW1sJykpLndyaXRlQXNTdHJpbmdTeW5jKAogICAgICAgICcnJwpuYW1lOiBleGFtcGxlCnZlcnNpb246IDAuMS4wCmVudmlyb25tZW50OgogIHNkazogXjIuMTcuMApkZXBlbmRlbmNpZXM6CiAgbWFzb246IGFueQogIGZvbzoKICAgIHBhdGg6IC4uLy4uL2ZvbwpkZXZfZGVwZW5kZW5jaWVzOgogIHRlc3Q6IGFueQogIGJhcjoKICAgIHBhdGg6IC4uLy4uL2JhcgonJycsCiAgICAgICk7CiAgICAgIGF3YWl0IGV4cGVjdExhdGVyKAogICAgICAgIHJlcG9ydEV4dGVybmFsUGF0aERlcGVuZGVuY2llcyhjb250ZXh0LCBkaXJlY3RvcnkpLAogICAgICAgIGNvbXBsZXRlcywKICAgICAgKTsKICAgICAgdmVyaWZ5KAogICAgICAgICgpID0+IGxvZ2dlci5lcnIoJ0FsbCBwYXRoIGRlcGVuZGVuY2llcyBtdXN0IGJlIHdpdGhpbiB0aGUgcHJvamVjdC4nKSwKICAgICAgKS5jYWxsZWQoMSk7CiAgICAgIHZlcmlmeSgKICAgICAgICAoKSA9PiBsb2dnZXIuZXJyKCdFeHRlcm5hbCBwYXRoIGRlcGVuZGVuY2llcyBkZXRlY3RlZDonKSwKICAgICAgKS5jYWxsZWQoMSk7CiAgICAgIHZlcmlmeSgKICAgICAgICAoKSA9PiBsb2dnZXIuZXJyKCcgIFx1ezIwMjJ9IGZvbyBmcm9tIC4uLy4uL2ZvbycpLAogICAgICApLmNhbGxlZCgxKTsKICAgICAgdmVyaWZ5KAogICAgICAgICgpID0+IGxvZ2dlci5lcnIoJyAgXHV7MjAyMn0gYmFyIGZyb20gLi4vLi4vYmFyJyksCiAgICAgICkuY2FsbGVkKDEpOwogICAgICBkaXJlY3RvcnkuZGVsZXRlKHJlY3Vyc2l2ZTogdHJ1ZSkuaWdub3JlKCk7CiAgICB9KTsKICB9KTsKfQo=",
+      "type": "text"
+    },
+    {
+      "path": "test/src/report_rogue_routes_test.dart",
+      "data":
+          "aW1wb3J0ICdwYWNrYWdlOmRhcnRfZnJvZ19nZW4vZGFydF9mcm9nX2dlbi5kYXJ0JzsKaW1wb3J0ICdwYWNrYWdlOm1hc29uL21hc29uLmRhcnQnOwppbXBvcnQgJ3BhY2thZ2U6bW9ja3RhaWwvbW9ja3RhaWwuZGFydCc7CmltcG9ydCAncGFja2FnZTp0ZXN0L3Rlc3QuZGFydCc7CgppbXBvcnQgJy4uLy4uL3NyYy9yZXBvcnRfcm9ndWVfcm91dGVzLmRhcnQnOwoKY2xhc3MgX01vY2tIb29rQ29udGV4dCBleHRlbmRzIE1vY2sgaW1wbGVtZW50cyBIb29rQ29udGV4dCB7fQoKY2xhc3MgX01vY2tMb2dnZXIgZXh0ZW5kcyBNb2NrIGltcGxlbWVudHMgTG9nZ2VyIHt9CgpjbGFzcyBfTW9ja1JvdXRlQ29uZmlndXJhdGlvbiBleHRlbmRzIE1vY2sgaW1wbGVtZW50cyBSb3V0ZUNvbmZpZ3VyYXRpb24ge30KCnZvaWQgbWFpbigpIHsKICBncm91cCgncmVwb3J0Um9ndWVSb3V0ZXMnLCAoKSB7CiAgICBsYXRlIEhvb2tDb250ZXh0IGNvbnRleHQ7CiAgICBsYXRlIExvZ2dlciBsb2dnZXI7CiAgICBsYXRlIFJvdXRlQ29uZmlndXJhdGlvbiBjb25maWd1cmF0aW9uOwoKICAgIHNldFVwKCgpIHsKICAgICAgY29udGV4dCA9IF9Nb2NrSG9va0NvbnRleHQoKTsKICAgICAgbG9nZ2VyID0gX01vY2tMb2dnZXIoKTsKICAgICAgY29uZmlndXJhdGlvbiA9IF9Nb2NrUm91dGVDb25maWd1cmF0aW9uKCk7CgogICAgICB3aGVuKCgpID0+IGNvbnRleHQubG9nZ2VyKS50aGVuUmV0dXJuKGxvZ2dlcik7CiAgICB9KTsKCiAgICB0ZXN0KCdyZXBvcnRzIG5vdGhpbmcgd2hlbiB0aGVyZSBhcmUgbm8gcm9ndWUgcm91dGVzJywgKCkgewogICAgICB3aGVuKCgpID0+IGNvbmZpZ3VyYXRpb24ucm9ndWVSb3V0ZXMpLnRoZW5SZXR1cm4oW10pOwogICAgICByZXBvcnRSb2d1ZVJvdXRlcyhjb250ZXh0LCBjb25maWd1cmF0aW9uKTsKICAgICAgdmVyaWZ5TmV2ZXIoKCkgPT4gbG9nZ2VyLmVycihhbnkoKSkpOwogICAgfSk7CgogICAgdGVzdCgncmVwb3J0cyBzaW5nbGUgcm9ndWUgcm91dGUnLCAoKSB7CiAgICAgIHdoZW4oKCkgPT4gY29uZmlndXJhdGlvbi5yb2d1ZVJvdXRlcykudGhlblJldHVybigKICAgICAgICBjb25zdCBbCiAgICAgICAgICBSb3V0ZUZpbGUoCiAgICAgICAgICAgIG5hbWU6ICdoZWxsbycsCiAgICAgICAgICAgIHBhdGg6ICdoZWxsby5kYXJ0JywKICAgICAgICAgICAgcm91dGU6ICcvaGVsbG8nLAogICAgICAgICAgICBwYXJhbXM6IFtdLAogICAgICAgICAgKSwKICAgICAgICBdLAogICAgICApOwogICAgICByZXBvcnRSb2d1ZVJvdXRlcyhjb250ZXh0LCBjb25maWd1cmF0aW9uKTsKICAgICAgdmVyaWZ5KAogICAgICAgICgpID0+IGxvZ2dlci5lcnIoCiAgICAgICAgICAnJydSb2d1ZSByb3V0ZSBkZXRlY3RlZC4ke2RlZmF1bHRGb3JlZ3JvdW5kLndyYXAoJyAnKX1SZW5hbWUgJHtsaWdodEN5YW4ud3JhcCgncm91dGVzL2hlbGxvLmRhcnQnKX0gdG8gJHtsaWdodEN5YW4ud3JhcCgncm91dGVzL2hlbGxvL2luZGV4LmRhcnQnKX0uJycnLAogICAgICAgICksCiAgICAgICk7CiAgICB9KTsKCiAgICB0ZXN0KCdyZXBvcnRzIG11bHRpcGxlIHJvZ3VlIHJvdXRlcycsICgpIHsKICAgICAgd2hlbigoKSA9PiBjb25maWd1cmF0aW9uLnJvZ3VlUm91dGVzKS50aGVuUmV0dXJuKAogICAgICAgIGNvbnN0IFsKICAgICAgICAgIFJvdXRlRmlsZSgKICAgICAgICAgICAgbmFtZTogJ2hlbGxvJywKICAgICAgICAgICAgcGF0aDogJ2hlbGxvLmRhcnQnLAogICAgICAgICAgICByb3V0ZTogJy9oZWxsbycsCiAgICAgICAgICAgIHBhcmFtczogW10sCiAgICAgICAgICApLAogICAgICAgICAgUm91dGVGaWxlKAogICAgICAgICAgICBuYW1lOiAnaGknLAogICAgICAgICAgICBwYXRoOiAnaGkuZGFydCcsCiAgICAgICAgICAgIHJvdXRlOiAnL2hpJywKICAgICAgICAgICAgcGFyYW1zOiBbXSwKICAgICAgICAgICksCiAgICAgICAgXSwKICAgICAgKTsKICAgICAgcmVwb3J0Um9ndWVSb3V0ZXMoY29udGV4dCwgY29uZmlndXJhdGlvbik7CiAgICAgIHZlcmlmeSgKICAgICAgICAoKSA9PiBsb2dnZXIuZXJyKAogICAgICAgICAgJycnUm9ndWUgcm91dGUgZGV0ZWN0ZWQuJHtkZWZhdWx0Rm9yZWdyb3VuZC53cmFwKCcgJyl9UmVuYW1lICR7bGlnaHRDeWFuLndyYXAoJ3JvdXRlcy9oZWxsby5kYXJ0Jyl9IHRvICR7bGlnaHRDeWFuLndyYXAoJ3JvdXRlcy9oZWxsby9pbmRleC5kYXJ0Jyl9LicnJywKICAgICAgICApLAogICAgICApOwogICAgICB2ZXJpZnkoCiAgICAgICAgKCkgPT4gbG9nZ2VyLmVycigKICAgICAgICAgICcnJ1JvZ3VlIHJvdXRlIGRldGVjdGVkLiR7ZGVmYXVsdEZvcmVncm91bmQud3JhcCgnICcpfVJlbmFtZSAke2xpZ2h0Q3lhbi53cmFwKCdyb3V0ZXMvaGkuZGFydCcpfSB0byAke2xpZ2h0Q3lhbi53cmFwKCdyb3V0ZXMvaGkvaW5kZXguZGFydCcpfS4nJycsCiAgICAgICAgKSwKICAgICAgKTsKICAgIH0pOwogIH0pOwp9Cg==",
+      "type": "text"
+    },
+    {
+      "path": "test/src/report_route_conflicts_test.dart",
+      "data":
+          "aW1wb3J0ICdwYWNrYWdlOmRhcnRfZnJvZ19nZW4vZGFydF9mcm9nX2dlbi5kYXJ0JzsKaW1wb3J0ICdwYWNrYWdlOm1hc29uL21hc29uLmRhcnQnOwppbXBvcnQgJ3BhY2thZ2U6bW9ja3RhaWwvbW9ja3RhaWwuZGFydCc7CmltcG9ydCAncGFja2FnZTp0ZXN0L3Rlc3QuZGFydCc7CgppbXBvcnQgJy4uLy4uL3NyYy9yZXBvcnRfcm91dGVfY29uZmxpY3RzLmRhcnQnOwoKY2xhc3MgX01vY2tIb29rQ29udGV4dCBleHRlbmRzIE1vY2sgaW1wbGVtZW50cyBIb29rQ29udGV4dCB7fQoKY2xhc3MgX01vY2tMb2dnZXIgZXh0ZW5kcyBNb2NrIGltcGxlbWVudHMgTG9nZ2VyIHt9CgpjbGFzcyBfTW9ja1JvdXRlQ29uZmlndXJhdGlvbiBleHRlbmRzIE1vY2sgaW1wbGVtZW50cyBSb3V0ZUNvbmZpZ3VyYXRpb24ge30KCnZvaWQgbWFpbigpIHsKICBncm91cCgncmVwb3J0Um91dGVDb25mbGljdHMnLCAoKSB7CiAgICBsYXRlIEhvb2tDb250ZXh0IGNvbnRleHQ7CiAgICBsYXRlIExvZ2dlciBsb2dnZXI7CiAgICBsYXRlIFJvdXRlQ29uZmlndXJhdGlvbiBjb25maWd1cmF0aW9uOwoKICAgIHNldFVwKCgpIHsKICAgICAgY29udGV4dCA9IF9Nb2NrSG9va0NvbnRleHQoKTsKICAgICAgbG9nZ2VyID0gX01vY2tMb2dnZXIoKTsKICAgICAgY29uZmlndXJhdGlvbiA9IF9Nb2NrUm91dGVDb25maWd1cmF0aW9uKCk7CgogICAgICB3aGVuKCgpID0+IGNvbnRleHQubG9nZ2VyKS50aGVuUmV0dXJuKGxvZ2dlcik7CiAgICB9KTsKCiAgICB0ZXN0KCdyZXBvcnRzIG5vdGhpbmcgd2hlbiB0aGVyZSBhcmUgbm8gZW5kcG9pbnRzJywgKCkgewogICAgICB3aGVuKCgpID0+IGNvbmZpZ3VyYXRpb24uZW5kcG9pbnRzKS50aGVuUmV0dXJuKHt9KTsKICAgICAgcmVwb3J0Um91dGVDb25mbGljdHMoY29udGV4dCwgY29uZmlndXJhdGlvbik7CiAgICAgIHZlcmlmeU5ldmVyKCgpID0+IGxvZ2dlci5lcnIoYW55KCkpKTsKICAgIH0pOwoKICAgIHRlc3QoJ3JlcG9ydHMgbm90aGluZyB3aGVuIHRoZXJlIGFyZSBlbmRwb2ludHMgYW5kIG5vIGNvbmZsaWN0cycsICgpIHsKICAgICAgd2hlbigoKSA9PiBjb25maWd1cmF0aW9uLmVuZHBvaW50cykudGhlblJldHVybih7CiAgICAgICAgJy8nOiBjb25zdCBbCiAgICAgICAgICBSb3V0ZUZpbGUoCiAgICAgICAgICAgIG5hbWU6ICdpbmRleCcsCiAgICAgICAgICAgIHBhdGg6ICdpbmRleC5kYXJ0JywKICAgICAgICAgICAgcm91dGU6ICcvJywKICAgICAgICAgICAgcGFyYW1zOiBbXSwKICAgICAgICAgICksCiAgICAgICAgXSwKICAgICAgICAnL2hlbGxvJzogY29uc3QgWwogICAgICAgICAgUm91dGVGaWxlKAogICAgICAgICAgICBuYW1lOiAnaGVsbG8nLAogICAgICAgICAgICBwYXRoOiAnaGVsbG8uZGFydCcsCiAgICAgICAgICAgIHJvdXRlOiAnL2hlbGxvJywKICAgICAgICAgICAgcGFyYW1zOiBbXSwKICAgICAgICAgICkKICAgICAgICBdCiAgICAgIH0pOwogICAgICByZXBvcnRSb3V0ZUNvbmZsaWN0cyhjb250ZXh0LCBjb25maWd1cmF0aW9uKTsKICAgICAgdmVyaWZ5TmV2ZXIoKCkgPT4gbG9nZ2VyLmVycihhbnkoKSkpOwogICAgfSk7CgogICAgdGVzdCgncmVwb3J0cyBzaW5nbGUgY29uZmxpY3Qgd2hlbiB0aGVyZSBpcyBvbmUgZW5kcG9pbnQgd2l0aCBjb25mbGljdHMnLAogICAgICAgICgpIHsKICAgICAgd2hlbigoKSA9PiBjb25maWd1cmF0aW9uLmVuZHBvaW50cykudGhlblJldHVybih7CiAgICAgICAgJy8nOiBjb25zdCBbCiAgICAgICAgICBSb3V0ZUZpbGUoCiAgICAgICAgICAgIG5hbWU6ICdpbmRleCcsCiAgICAgICAgICAgIHBhdGg6ICdpbmRleC5kYXJ0JywKICAgICAgICAgICAgcm91dGU6ICcvJywKICAgICAgICAgICAgcGFyYW1zOiBbXSwKICAgICAgICAgICksCiAgICAgICAgXSwKICAgICAgICAnL2hlbGxvJzogY29uc3QgWwogICAgICAgICAgUm91dGVGaWxlKAogICAgICAgICAgICBuYW1lOiAnaGVsbG8nLAogICAgICAgICAgICBwYXRoOiAnaGVsbG8uZGFydCcsCiAgICAgICAgICAgIHJvdXRlOiAnL2hlbGxvJywKICAgICAgICAgICAgcGFyYW1zOiBbXSwKICAgICAgICAgICksCiAgICAgICAgICBSb3V0ZUZpbGUoCiAgICAgICAgICAgIG5hbWU6ICdoZWxsb19pbmRleCcsCiAgICAgICAgICAgIHBhdGg6ICdoZWxsby9pbmRleC5kYXJ0JywKICAgICAgICAgICAgcm91dGU6ICcvJywKICAgICAgICAgICAgcGFyYW1zOiBbXSwKICAgICAgICAgICkKICAgICAgICBdCiAgICAgIH0pOwogICAgICByZXBvcnRSb3V0ZUNvbmZsaWN0cyhjb250ZXh0LCBjb25maWd1cmF0aW9uKTsKICAgICAgdmVyaWZ5KAogICAgICAgICgpID0+IGxvZ2dlci5lcnIoCiAgICAgICAgICAnJydSb3V0ZSBjb25mbGljdCBkZXRlY3RlZC4gJHtsaWdodEN5YW4ud3JhcCgncm91dGVzL2hlbGxvLmRhcnQnKX0gYW5kICR7bGlnaHRDeWFuLndyYXAoJ3JvdXRlcy9oZWxsby9pbmRleC5kYXJ0Jyl9IGJvdGggcmVzb2x2ZSB0byAke2xpZ2h0Q3lhbi53cmFwKCcvaGVsbG8nKX0uJycnLAogICAgICAgICksCiAgICAgICk7CiAgICB9KTsKCiAgICB0ZXN0KAogICAgICAgICdyZXBvcnRzIG11bHRpcGxlIGNvbmZsaWN0cyAnCiAgICAgICAgJ3doZW4gdGhlcmUgYXJlIG11bHRpcGxlIGVuZHBvaW50IHdpdGggY29uZmxpY3RzJywgKCkgewogICAgICB3aGVuKCgpID0+IGNvbmZpZ3VyYXRpb24uZW5kcG9pbnRzKS50aGVuUmV0dXJuKHsKICAgICAgICAnLyc6IGNvbnN0IFsKICAgICAgICAgIFJvdXRlRmlsZSgKICAgICAgICAgICAgbmFtZTogJ2luZGV4JywKICAgICAgICAgICAgcGF0aDogJ2luZGV4LmRhcnQnLAogICAgICAgICAgICByb3V0ZTogJy8nLAogICAgICAgICAgICBwYXJhbXM6IFtdLAogICAgICAgICAgKSwKICAgICAgICBdLAogICAgICAgICcvaGVsbG8nOiBjb25zdCBbCiAgICAgICAgICBSb3V0ZUZpbGUoCiAgICAgICAgICAgIG5hbWU6ICdoZWxsbycsCiAgICAgICAgICAgIHBhdGg6ICdoZWxsby5kYXJ0JywKICAgICAgICAgICAgcm91dGU6ICcvaGVsbG8nLAogICAgICAgICAgICBwYXJhbXM6IFtdLAogICAgICAgICAgKSwKICAgICAgICAgIFJvdXRlRmlsZSgKICAgICAgICAgICAgbmFtZTogJ2hlbGxvX2luZGV4JywKICAgICAgICAgICAgcGF0aDogJ2hlbGxvL2luZGV4LmRhcnQnLAogICAgICAgICAgICByb3V0ZTogJy8nLAogICAgICAgICAgICBwYXJhbXM6IFtdLAogICAgICAgICAgKQogICAgICAgIF0sCiAgICAgICAgJy9lY2hvJzogY29uc3QgWwogICAgICAgICAgUm91dGVGaWxlKAogICAgICAgICAgICBuYW1lOiAnZWNobycsCiAgICAgICAgICAgIHBhdGg6ICdlY2hvLmRhcnQnLAogICAgICAgICAgICByb3V0ZTogJy9lY2hvJywKICAgICAgICAgICAgcGFyYW1zOiBbXSwKICAgICAgICAgICksCiAgICAgICAgICBSb3V0ZUZpbGUoCiAgICAgICAgICAgIG5hbWU6ICdlY2hvX2luZGV4JywKICAgICAgICAgICAgcGF0aDogJ2VjaG8vaW5kZXguZGFydCcsCiAgICAgICAgICAgIHJvdXRlOiAnLycsCiAgICAgICAgICAgIHBhcmFtczogW10sCiAgICAgICAgICApCiAgICAgICAgXQogICAgICB9KTsKICAgICAgcmVwb3J0Um91dGVDb25mbGljdHMoY29udGV4dCwgY29uZmlndXJhdGlvbik7CiAgICAgIHZlcmlmeSgKICAgICAgICAoKSA9PiBsb2dnZXIuZXJyKAogICAgICAgICAgJycnUm91dGUgY29uZmxpY3QgZGV0ZWN0ZWQuICR7bGlnaHRDeWFuLndyYXAoJ3JvdXRlcy9oZWxsby5kYXJ0Jyl9IGFuZCAke2xpZ2h0Q3lhbi53cmFwKCdyb3V0ZXMvaGVsbG8vaW5kZXguZGFydCcpfSBib3RoIHJlc29sdmUgdG8gJHtsaWdodEN5YW4ud3JhcCgnL2hlbGxvJyl9LicnJywKICAgICAgICApLAogICAgICApOwogICAgICB2ZXJpZnkoCiAgICAgICAgKCkgPT4gbG9nZ2VyLmVycigKICAgICAgICAgICcnJ1JvdXRlIGNvbmZsaWN0IGRldGVjdGVkLiAke2xpZ2h0Q3lhbi53cmFwKCdyb3V0ZXMvZWNoby5kYXJ0Jyl9IGFuZCAke2xpZ2h0Q3lhbi53cmFwKCdyb3V0ZXMvZWNoby9pbmRleC5kYXJ0Jyl9IGJvdGggcmVzb2x2ZSB0byAke2xpZ2h0Q3lhbi53cmFwKCcvZWNobycpfS4nJycsCiAgICAgICAgKSwKICAgICAgKTsKICAgIH0pOwogIH0pOwp9Cg==",
       "type": "text"
     }
   ],
